refactor: 로또 생성 로직 리팩토링
1. List<LottoNumber> 로직 제거 -> Set<LottoNumber> 로 대체
(LottoTicket 생성 로직에 List<Integer> -> List<LottoNumber> -> Set<LottoNumber> 로직 불필요)

2. LottoGenerator 삭제
